fix: Se asignan bien las vistas y se cambian los update
