+++ Controllers/LoginController.cs
@@ -5,10 +5,22 @@ using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
-namespace Library.Controllers
-{
+using Library.DataBase;
+using Library.Models;
+using Microsoft.EntityFrameworkCore;
+namespace Library.Controllers;
+
+
     public class LoginController : Controller
     {
+        private readonly AplicationDbContext _context;
+
+        public LoginController(AplicationDbContext context)
+        {
+            _context = context;
+        }
+
+
         // GET: /Login/Index
         [HttpGet]
         public IActionResult Index()
@@ -18,7 +30,8 @@ namespace Library.Controllers
         [HttpGet]
         public IActionResult Dashboard()
         {
-            return View();
+            var user = _context.Users.ToList();
+            return View(user);
         }
 
         // POST: /Login/Index
@@ -43,5 +56,41 @@ namespace Library.Controllers
         {
             return View();
         }
+
+        [HttpGet]
+        public async Task<IActionResult> Update(int id)
+        {
+            User user = await _context.Users.FirstAsync(u => u.Id == id);
+            return View(user);
+        }
+
+        [HttpPost]
+        public async Task<IActionResult> Update(User user)
+        {
+            _context.Users.Update(user);
+            await _context.SaveChangesAsync();
+            return RedirectToAction(nameof(Dashboard));
+        }
+
+        [HttpGet]
+        public async Task<IActionResult> Delete(int id)
+        {
+            var user = await _context.Users.FindAsync(id);
+            if (user == null)
+            {
+                return NotFound("User not found");
+            }
+
+            _context.Users.Remove(user);
+            await _context.SaveChangesAsync();
+            return RedirectToAction(nameof(Dashboard));
+        }
+
+
+        [HttpGet]
+        public async Task<IActionResult> List()
+        {
+            List<User> list = await _context.Users.ToListAsync();
+            return View(list);
+        }
     }
-}
