--- Views/User/Dashboard.cshtml
@@ -1,53 +1,0 @@
-@model List<Library.Models.User>;
-
-<div class="text-center">
-    <h1 class="display-4">¡Estos son los Usuarios Registrados!</h1><br>
-
-    @if (Model != null && Model.Any())
-    {
-        <div class="table-responsive">
-            <table class="table table-bordered table-striped">
-                <thead>
-                    <tr>
-                        <th>Id</th>
-                        <th>Nombre Completo</th>
-                        <th>Dirección</th>
-                        <th>Teléfono</th>
-                        <th>Número de Documento</th>
-                        <th>Rol</th>
-                        <th>Acciones</th>
-                    </tr>
-                </thead>
-                <tbody>
-                    @foreach (var user in Model)
-                    {
-                        <tr>
-                            <td>@user.Id</td>
-                            <td>@user.Name @user.LastName</td>
-                            <td>@user.Address</td>
-                            <td>@user.PhoneNumber</td>
-                            <td>@user.NumberDocument</td>
-                            <td>@user.Role</td>
-                            <td>
-                                <a class="btn btn-warning btn-sm" asp-controller="User" asp-action="UpdateUser" asp-route-id="@user.Id">Editar</a>
-                                <a class="btn btn-danger btn-sm" asp-controller="User" asp-action="DeleteUser" asp-route-id="@user.Id" onclick="return confirmar()">Eliminar</a>
-                            </td>
-                        </tr>
-                    }
-                </tbody>
-            </table>
-        </div>
-    }
-    else
-    {
-        <p>No hay usuarios para mostrar.</p>
-    }
-
-@section Scripts {
-    <script>
-        function confirmar() {
-            return confirm("¿Estás seguro de Eliminar este user?");
-        }
-    </script>
-}
-
